ADD CurrencyExchangeHeader
