+++ lib/core/utils/app_colors.dart
@@ -5,6 +5,7 @@ abstract class AppColors {
   static Color primaryColor = const Color(0xff131316);
   static Color secondaryColor = const Color(0xff212126);
   static Color appBlueColor = CupertinoColors.activeBlue;
+  static Color appGreenColor = CupertinoColors.activeGreen;
   static Color appBlueSecondaryColor = const Color(0xff001664);
   static Color hint = Colors.grey;
